LPD-5780 - merge commerce products page implementations
